Kdump improvements
- Added support for reporting kdump reboot cause to detect and report
a kernel crash
- Added additional commands to sonic-kdump-config wrapper script to
dump Kdump configuration and status information in JSON format
- Automatically save Kdump configuration in startup configuration file.
This is an added convenience to the user as Kdump configuration changes
require a system reboot and user is expected to save the configuration so
that they are activated post reboot.
- Additional bug fixes and improvements to allow hostcfgd to process Kdump
configuration changes in ConfigDB. sonic-kdump-config script will write to
ConfigDB.
- Moved kdump CLI commands to a dedicated file
