+++ show/main.py
@@ -20,6 +20,7 @@ from . import bgp_common
 from . import chassis_modules
 from . import feature
 from . import fgnhg
+from . import kdump
 from . import interfaces
 from . import kube
 from . import mlnx
@@ -131,6 +132,7 @@ def cli(ctx):
 cli.add_command(chassis_modules.chassis_modules)
 cli.add_command(feature.feature)
 cli.add_command(fgnhg.fgnhg)
+cli.add_command(kdump.kdump)
 cli.add_command(interfaces.interfaces)
 cli.add_command(kube.kubernetes)
 cli.add_command(muxcable.muxcable)
@@ -1422,87 +1424,6 @@ def system_memory(verbose):
     """Show memory information"""
     cmd = "free -m"
     run_command(cmd, display_cmd=verbose)
-
-#
-# 'kdump command ("show kdump ...")
-#
-@cli.group(cls=clicommon.AliasedGroup)
-def kdump():
-    """Show kdump configuration, status and information """
-    pass
-
-@kdump.command('enabled')
-def enabled():
-    """Show if kdump is enabled or disabled"""
-    kdump_is_enabled = False
-    config_db = ConfigDBConnector()
-    if config_db is not None:
-        config_db.connect()
-        table_data = config_db.get_table('KDUMP')
-        if table_data is not None:
-            config_data = table_data.get('config')
-            if config_data is not None:
-                if config_data.get('enabled').lower() == 'true':
-                    kdump_is_enabled = True
-    if kdump_is_enabled:
-        click.echo("kdump is enabled")
-    else:
-        click.echo("kdump is disabled")
-
-@kdump.command('status')
-def status():
-    """Show kdump status"""
-    run_command("sonic-kdump-config --status")
-    run_command("sonic-kdump-config --memory")
-    run_command("sonic-kdump-config --num_dumps")
-    run_command("sonic-kdump-config --files")
-
-@kdump.command('memory')
-def memory():
-    """Show kdump memory information"""
-    kdump_memory = "0M-2G:256M,2G-4G:320M,4G-8G:384M,8G-:448M"
-    config_db = ConfigDBConnector()
-    if config_db is not None:
-        config_db.connect()
-        table_data = config_db.get_table('KDUMP')
-        if table_data is not None:
-            config_data = table_data.get('config')
-            if config_data is not None:
-                kdump_memory_from_db = config_data.get('memory')
-                if kdump_memory_from_db is not None:
-                    kdump_memory = kdump_memory_from_db
-    click.echo("Memory Reserved: %s" % kdump_memory)
-
-@kdump.command('num_dumps')
-def num_dumps():
-    """Show kdump max number of dump files"""
-    kdump_num_dumps = "3"
-    config_db = ConfigDBConnector()
-    if config_db is not None:
-        config_db.connect()
-        table_data = config_db.get_table('KDUMP')
-        if table_data is not None:
-            config_data = table_data.get('config')
-            if config_data is not None:
-                kdump_num_dumps_from_db = config_data.get('num_dumps')
-                if kdump_num_dumps_from_db is not None:
-                    kdump_num_dumps = kdump_num_dumps_from_db
-    click.echo("Maximum number of Kernel Core files Stored: %s" % kdump_num_dumps)
-
-@kdump.command('files')
-def files():
-    """Show kdump kernel core dump files"""
-    run_command("sonic-kdump-config --files")
-
-@kdump.command()
-@click.argument('record', required=True)
-@click.argument('lines', metavar='<lines>', required=False)
-def log(record, lines):
-    """Show kdump kernel core dump file kernel log"""
-    if lines is None:
-        run_command("sonic-kdump-config --file %s" % record)
-    else:
-        run_command("sonic-kdump-config --file %s --lines %s" % (record, lines))
 
 @cli.command('services')
 def services():
